# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MeepShop Chat API",
    "version": "1.0.0",
...
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string",
            "description": "Stack trace (only in development)"
          }
        }
      },
...
      "MessageAuthor": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d0fe4f5311236168a109ca"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "legacyUserId": {
            "type": "integer",
            "example": 101,
            "nullable": true
          }
        }
      },
      "Reaction": {
        "type": "object",
        "description": "Key-value pairs of reaction emoji and count.",
        "additionalProperties": {
          "type": "integer",
          "example": 1
        },
        "example": {
          "👍": 2,
          "❤️": 1
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d0fe4f5311236168a109cd"
          },
          "conversationId": {
            "type": "string",
            "example": "60d0fe4f5311236168a109ce"
          },
          "senderId": {
            "$ref": "#/components/schemas/MessageAuthor"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          },
          "content": {
            "type": "string",
            "example": "Hello world!"
          },
          "reactions": {
            "$ref": "#/components/schemas/Reaction"
          },
          "legacyConvId": {
            "type": "integer",
            "example": 1,
            "nullable": true
          },
          "legacySenderId": {
            "type": "integer",
            "example": 101,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateMessageInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image"
            ],
            "example": "text"
          },
          "content": {
            "type": "string",
            "example": "This is a test message. For image type, this will be the S3 URL."
          }
        },
        "required": [
          "type",
          "content"
        ]
      },
...
    }
  },
  "paths": {
...
    "/conversations/{conversationId}/messages": {
      "post": {
        "summary": "Create a new message in a specific conversation",
        "tags": [
          "Conversations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the conversation where the message will be posted"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "$ref": "#/components/schemas/MessageResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing type or content)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (user not part of conversation or not logged in)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/messages": {
      "get": {
        "summary": "Get messages for a specific conversation with pagination",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "conversationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the conversation to fetch messages for."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of messages per page."
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt:asc"
            },
            "description": "Sort order for messages (e.g., 'createdAt:asc' or 'createdAt:desc')."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MessageResponse"
                          }
                        },
                        "currentPage": {
                          "type": "integer",
                          "example": 1
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 3
                        },
                        "totalMessages": {
                          "type": "integer",
                          "example": 150
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing conversationId)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (user not part of conversation or not logged in)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}